pyproject.toml
[project]
name = "telegrambot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "python-telegram-bot>=22.5",
    "pytz>=2025.2",
    "requests>=2.32.5",
]
